requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
gspread>=6.0
google-auth>=2.25
jinja2>=3.1
python-dotenv>=1.0
pytest>=8.0
